+++ src/index.html
@@ -6,6 +6,9 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <meta http-equiv="X-UA-Compatible" content="ie=edge" />
   <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="./pages/login/login-style.css">
+  <link rel="stylesheet" href="./pages/cadastro/cadastro-style.css">
+  <link rel="stylesheet" href="./pages/perfil/perfil-style.css">
   <link rel="stylesheet" href=" https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css" />
   <title>Document</title>
 </head>
@@ -31,60 +34,7 @@
       </nav>
     </header>
 
-
-
-
     <main id="conteudo-principal"></main>
-    <!-- 
-    <nav>
-      <ul>
-        <li>
-          <a href="#">Home</a>
-        </li>
-        <li>
-          <a href="#">Login</a>
-        </li>
-        <li>
-          <a href="#">Contato</a>
-        </li>
-      </ul>
-    </nav>
-
-  </div>
-
-<div id="banner-section">
-  <img id="banner" src="img/banner.jpg" alt="Foto demonstrativo do site">
-  <div id="texto-sobre">
-    <h2>Fronters</h2>
-    <h3>De frente com o Front</h3>
-    <h4>Está preparado para invadir 
-      o mundo dos Devs? Venha trocar experiências, aprender, ensinar e conhecer novas pessoas</h4>
-
-  </div>
-</div>
-
-
-
-  <form class="campo-form">
-    <div>
-      <label for="emailUsuario">Email:</label>
-      <input type="email" id="emailUsuario" placeholder="Digite o seu email">
-    </div>
-    <div>
-      <label for="senhaUsuario">Senha:</label>
-      <input type="password" id="senhaUsuario" placeholder="Digite a sua senha">
-    </div>
-    <div>
-      <button type="button" id="botaoCadastro">Cadastro</button>
-    </div>
-    <div>
-      <button type="button" id="botaoLogin">Login</button>
-    </div>
-    <div>
-      <button id="botaoGoogle">Entrar com a conta Google</button>
-    </div>
-  </form>
- -->
 
     <footer id="footer">
       <h1 class="logo">FRONTERS</h1>
@@ -95,13 +45,7 @@
         <a href="https://github.com/joiceesouza" target="_blank">Joice<i class="fab fa-github"></i></a>
       </div>
 
-
     </footer>
-
-
-
-
-
