+++ src/main.js
@@ -1,36 +1,43 @@
 // Este é o ponto de entrada da sua aplicação
 
 // import { testeCadastro } from './lib/index.js';
-import login from './pages/login/index.js';
-import cadastro from './pages/cadastro/index.js';
+import {TemplateLogin, AddEventoLogin, AddEventoLoginComGoogle} from './pages/login/index.js';
+import { TemplateCadastro, AddEventoCadastro, PopUpCadastro } from './pages/cadastro/index.js';
 import perfil from './pages/perfil/index.js';
 
 const main = document.querySelector('#conteudo-principal');
 const init = () => {
-    window.addEventListener("hashchange", () => {
-        main.innerHTML = "";
-        switch (window.location.hash) {
-            case " ":
-                main.appendChild(login())
-                break;
-            case "#perfil":
-                main.appendChild(perfil())
-                break;
-            case "#cadastro":
-                main.appendChild(cadastro())
-                break;
-            default:
-                main.appendChild(login())
+    carregarPaginaURL();
+    window.addEventListener("hashchange", carregarPaginaURL);
+}
 
-
-        }
-    })
+function carregarPaginaURL() {
+    main.innerHTML = "";
+    switch (window.location.hash) {
+        case " ":
+            main.appendChild(TemplateLogin())
+            AddEventoLogin();
+            AddEventoLoginComGoogle();
+            break;
+        case "#perfil":
+            main.appendChild(perfil())
+            break;
+        case "#cadastro":
+            main.appendChild(TemplateCadastro())
+            AddEventoCadastro();
+            PopUpCadastro();
+            break;
+        default:
+            main.appendChild(TemplateLogin())
+            AddEventoLogin();
+            AddEventoLoginComGoogle();
+    }
 }
 
 
 
 window.addEventListener("load", () => {
-    main.appendChild(login());
+    //main.appendChild(login());
     init();
 
 });
@@ -46,98 +53,14 @@ window.addEventListener("load", () => {
 // })
 
 
-const botaoDoCadastro = document.querySelector('#botaoFinalizarCadastro');
-
-botaoDoCadastro.addEventListener("click", () => {
-    const email = document.getElementById('emailUsuarioCadastro').value;
-    const password = document.getElementById('confirmaSenhaUsuario').value;
-
-    firebase
-        .auth()
-        .createUserWithEmailAndPassword(email, password)
-        .then((userCredential) => {
-            // Signed in
-            const user = userCredential.user;
-            console.log(user, 'Cadastrado!');
-            // ...
-        })
-        .catch((error) => {
-            const errorCode = error.code;
-            const errorMessage = error.message;
-            console.log('Infelizmente aconteceu algum erro!', errorCode, errorMessage);
-            // ..
-        });
-})
-
-
-
-// //CONECTAR USUARIO com endereço de e-mail e senha
-
-const botaoDoLogin = document.getElementById('botaoLogin');
-botaoDoLogin.addEventListener("click", () => {
-
-    const email = document.getElementById('emailUsuario').value;
-    const password = document.getElementById('senhaUsuario').value;
-
-    firebase
-        .auth()
-        .signInWithEmailAndPassword(email, password)
-        .then((userCredential) => {
-            // Signed in
-            //window.location = "pagina que ira depois q fazer login"
-            const user = userCredential.user;
-            console.log(user, 'Deu certo o login! ihull');
-            // ...
-        })
-        .catch((error) => {
-            const errorCode = error.code;
-            const errorMessage = error.message;
-            console.log('Infelizmente aconteceu algum com login!', errorCode, errorMessage);
-        });
-
-});
-
-
-
-
-//   //LOGIN GOOGLE
-const botaoDoGoogle = document.getElementById('botaoGoogle')
-botaoDoGoogle.addEventListener('click', (event) => {
-    alert(1)
-    event.preventDefault()
-    let provider = new firebase.auth.GoogleAuthProvider();
-
-    //Para fazer login com uma janela pop-up, chame signInWithPopup
-    firebase
-        .auth()
-        .signInWithPopup(provider)
-        .then((result) => {
-            console.log(result)
-            alert(1)
-            /** @type {firebase.auth.OAuthCredential} */
-            let credential = result.credential;
-
-            // This gives you a Google Access Token. You can use it to access the Google API.
-            let token = credential.accessToken;
-            // The signed-in user info.
-            let user = result.user;
-            // ...
-        }).catch((error) => {
-            // Handle Errors here.
-            let errorCode = error.code;
-            let errorMessage = error.message;
-            // The email of the user's account used.
-            let email = error.email;
-            // The firebase.auth.AuthCredential type that was used.
-            let credential = error.credential;
-            // ...
-        });
 
 
 
 
 
-})
+
+
+
