+++ src/pages/login/login-style.css
@@ -1,0 +1,82 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    text-decoration: none;
+}
+
+
+:root {
+    --bg-color: #6ab1d1;
+    --rodape-footer: #002855;
+    --logo: #fb8500;
+  }
+
+
+
+#texto-sobre {
+    text-align: center;
+    margin: 1em;
+}
+
+form {
+    margin-bottom: 5em;
+}
+
+label {
+    display: block;
+}
+#emailUsuario, #senhaUsuario {
+    width: 80vw;
+    background-color: white;
+    border: solid 2px black;
+
+}
+
+#botaoCadastrar a{
+    color: rgb(7, 7, 7);
+}
+
+#banner {
+    width: 100%;
+    height: 30vh;
+}
+
+
+.campo-form {
+    margin: 20px;
+    width: 100%;
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+#footer {
+    width: 100vw;
+    height: 20vh;
+    background-color: var(--rodape-footer);
+    text-align: center;
+}
+
+
+#footer h1 {
+    font-size: 20px;
+}
+
+#github a {
+    color: white;
+}
+
+#github i {
+    color: #fb8500;
+    margin-left: 10px;
+}
