+++ src/style.css
@@ -91,28 +91,6 @@ margin: 8px;
 }
 
 
-
-
-
-
-
-
-/* .one,
-.two,
-.three {
-      background-color: #fff;
-      width: 4em;
-      height: 5px;
-      margin: 6px auto;
-      transition-duration: 0.3s;
-} */
-
-
-
-
-
-
-
 /* .container-header {
     display: flex;
     justify-content: space-evenly;
@@ -120,50 +98,6 @@ margin: 8px;
     /* width: 100vh; 
     /* height: 8vh; 
 } */
-
-
-
-#texto-sobre {
-    text-align: center;
-    margin: 1em;
-}
-
-form {
-    margin-bottom: 5em;
-}
-
-label {
-    display: block;
-}
-#emailUsuario, #senhaUsuario {
-    width: 80vw;
-    background-color: white;
-    border: solid 2px black;
-
-}
-
-#botaoCadastrar a{
-    color: rgb(7, 7, 7);
-}
-
-
-
-
-
-
-
-#banner {
-    width: 100%;
-    height: 30vh;
-}
-
-
-.campo-form {
-    margin: 20px;
-    width: 100%;
-}
-
-
 
 
 #footer {
@@ -187,6 +121,7 @@ label {
     margin-left: 10px;
 }
 
+
 @media screen and (min-width: 601px){
      #menu-hamburguer {
          display: none;
@@ -202,6 +137,9 @@ label {
         margin-left: 32px;
     }
 
+    .popup {
+        max-width: 300px;
+    }
 
 }
